avance V2.8

+++ src/controllers/ProjectController.ts
@@ -19,7 +19,11 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    {manager: {$in: req.user._id}}
+                ]
+            })
             res.json(projects)
         } catch (error) {
             console.log(error)
@@ -32,6 +36,11 @@ export class ProjectController {
             const project = await Project.findById(id).populate('tasks')
             if(!project){
                 const error = new Error('Proyecto no encontrado')
+                return res.status(404).json({error: error.message})
+            }
+
+            if(project.manager.toString() !== req.user._id.toString() ){
+                const error = new Error('Accion no valida')
                 return res.status(404).json({error: error.message})
             }
         res.json(project)
@@ -48,6 +57,12 @@ export class ProjectController {
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({error: error.message})
             }
+
+            if(project.manager.toString() !== req.user._id.toString() ){
+                const error = new Error('Solo el Manager úede actualizar un Proyecto')
+                return res.status(404).json({error: error.message})
+            }
+
             project.clientName = req.body.clientName
             project.projectName = req.body.projectName
             project.description = req.body.description
@@ -67,6 +82,12 @@ export class ProjectController {
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({error: error.message})
             }
+
+            if(project.manager.toString() !== req.user._id.toString() ){
+                const error = new Error('Solo el manager puede eliminar un Proyecto')
+                return res.status(404).json({error: error.message})
+            }
+
             await project.deleteOne()
             res.send('Proyecto Eliminado')            
         } catch (error) {
